Fix count down
