+++ convex/races.ts
@@ -133,6 +133,8 @@ export const getPlayerRaces = query({
       }
     }
 
+    console.log({ len: playerRacesWithPlayers.length });
+
     return playerRacesWithPlayers;
   },
 });
@@ -145,6 +147,8 @@ export const joinRace = mutation({
   handler: async (ctx, args) => {
     const userID = await getUserFromToken(args.token);
     requireAuth(userID);
+
+    console.log({ userID });
 
     // Check if player is already in this race
     const existing = await ctx.db
