--- src/routes/race/RaceArea.tsx
@@ -1,47 +1,0 @@
-import { onMount, Show } from "solid-js";
-import { createMutation } from "../../convex-solid";
-import { api } from "../../../convex/_generated/api";
-import { getCurrentUser } from "../../convex";
-import { Race, PlayerRaceWithPlayer } from "../../types";
-import { RaceInput } from "./RaceInput";
-
-export function RaceArea(props: {
-  race: Race;
-  playerRaces: PlayerRaceWithPlayer[];
-  quote: string;
-}) {
-  const { userID } = getCurrentUser();
-
-  function playerRace() {
-    return (props.playerRaces ?? []).find((r) => r.playerID === userID);
-  }
-
-  return (
-    <Show when={playerRace()} fallback={<Initializer race={props.race} />}>
-      {(playerRace) => (
-        <RaceInput
-          race={props.race}
-          playerRace={playerRace()}
-          playerRaces={props.playerRaces}
-          quote={props.quote}
-        />
-      )}
-    </Show>
-  );
-}
-
-function Initializer(props: { race: Race }) {
-  const { token } = getCurrentUser();
-  const joinRace = createMutation(api.races.joinRace);
-
-  // Initialize player race on mount
-  onMount(async () => {
-    if (["ready", "started", "starting"].includes(props.race.status)) {
-      await joinRace({
-        raceId: props.race._id,
-        ...(token ? { token } : {}),
-      });
-    }
-  });
-  return <div />;
-}
